v0.0.14 - Basic Screenshake Implementation
Very very basic implementation of screen shake using both a perlin noise and a sin wave approach.

Sin wave works and feels a ton better sue to predictable movement that is easy for your eyes to follow (and it returns to zero better/smoother).

p and l keys to test them respectively.